Added the ability to take a server URL.
This was mainly to make it work with wordpress, as it tends to put a
canonical URL in front of all the resources.

I'm not sure if we should take *all* the ServerName and ServerAlias of
Apache into account, or just the url used to make the call.

So as it is, if someone hits your site by putting http://supa.ws in the
browser:

 * /images/x.gif == http://supa.ws/images/x.gif

.. but ..

 * /images/x.gif != http://www.supa.ws/images/x.gif

The way to get around that is to to explicitly add:

CDN_URL /images/ images.cdn.supa.ws/
CDN_URL http://www.supa.ws/images/ images.cdn.supa.ws/

...

I think the next step is to make it understand ServerName and
ServerAlias, and that that should be done through 'lower_bound' and a
sorted vector.

However I'm not 100% convinced it'll be a useful feature yet.